combined the insert and set base classes to prevent code dupe
