+++ dbflow/src/main/java/com/raizlabs/android/dbflow/sql/language/Insert.java
@@ -1,30 +1,21 @@
 package com.raizlabs.android.dbflow.sql.language;
 
 import android.content.ContentValues;
-import android.database.Cursor;
 
 import com.raizlabs.android.dbflow.annotation.ConflictAction;
 import com.raizlabs.android.dbflow.config.FlowManager;
-import com.raizlabs.android.dbflow.sql.Query;
-import com.raizlabs.android.dbflow.sql.SqlUtils;
 import com.raizlabs.android.dbflow.sql.builder.ValueQueryBuilder;
 import com.raizlabs.android.dbflow.sql.language.property.IProperty;
-import com.raizlabs.android.dbflow.sql.queriable.Queriable;
 import com.raizlabs.android.dbflow.structure.Model;
 import com.raizlabs.android.dbflow.structure.ModelAdapter;
-import com.raizlabs.android.dbflow.structure.database.DatabaseWrapper;
 
 import java.util.Map;
 
 /**
  * Description: The SQLite INSERT command
  */
-public class Insert<ModelClass extends Model> implements Query, Queriable {
+public class Insert<ModelClass extends Model> extends BaseQueriable<ModelClass> {
 
-    /**
-     * The table class that this INSERT points to
-     */
-    private Class<ModelClass> table;
 
     /**
      * The columns to specify in this query (optional)
@@ -47,7 +38,7 @@ public class Insert<ModelClass extends Model> implements Query, Queriable {
      * @param table The table to insert into
      */
     public Insert(Class<ModelClass> table) {
-        this.table = table;
+        super(table);
     }
 
     /**
@@ -55,11 +46,10 @@ public class Insert<ModelClass extends Model> implements Query, Queriable {
      * each column has a 1-1 relationship to the values.
      *
      * @param columns The columns to use
-     * @return This INSERT statement
      */
     public Insert<ModelClass> columns(String... columns) {
         this.columns = new IProperty[columns.length];
-        ModelAdapter<ModelClass> modelClassModelAdapter = FlowManager.getModelAdapter(table);
+        ModelAdapter<ModelClass> modelClassModelAdapter = FlowManager.getModelAdapter(getTable());
         for (int i = 0; i < columns.length; i++) {
             String column = columns[i];
             this.columns[i] = modelClassModelAdapter.getProperty(column);
@@ -72,7 +62,6 @@ public class Insert<ModelClass extends Model> implements Query, Queriable {
      * a set of columns are specified.
      *
      * @param values The non type-converted values
-     * @return
      */
     public Insert<ModelClass> values(Object... values) {
         this.values = values;
@@ -83,7 +72,6 @@ public class Insert<ModelClass extends Model> implements Query, Queriable {
      * Uses the {@link Condition} pairs to fill this insert query.
      *
      * @param conditions The conditions that we use to fill the columns and values of this INSERT
-     * @return
      */
     public Insert<ModelClass> columnValues(SQLCondition... conditions) {
 
@@ -102,8 +90,7 @@ public class Insert<ModelClass extends Model> implements Query, Queriable {
     /**
      * Uses the {@link Condition} pairs to fill this insert query.
      *
-     * @param conditionQueryBuilder The condition query builder to use
-     * @return
+     * @param conditionGroup The ConditionGroup to use
      */
     public Insert<ModelClass> columnValues(ConditionGroup conditionGroup) {
 
@@ -194,19 +181,6 @@ public class Insert<ModelClass extends Model> implements Query, Queriable {
         return or(ConflictAction.IGNORE);
     }
 
-    /**
-     * @return Exeuctes and returns the count of rows affected by this query.
-     */
-    @Override
-    public long count(DatabaseWrapper databaseWrapper) {
-        return SqlUtils.longForQuery(databaseWrapper, getQuery());
-    }
-
-    @Override
-    public long count() {
-        return count(FlowManager.getDatabaseForTable(table).getWritableDatabase());
-    }
-
     @Override
     public String getQuery() {
         ValueQueryBuilder queryBuilder = new ValueQueryBuilder("INSERT ");
@@ -214,7 +188,7 @@ public class Insert<ModelClass extends Model> implements Query, Queriable {
             queryBuilder.append("OR ").append(conflictAction);
         }
         queryBuilder.appendSpaceSeparated("INTO")
-                .appendTableName(table);
+                .appendTableName(getTable());
 
         if (columns != null) {
             queryBuilder.append("(")
@@ -223,10 +197,10 @@ public class Insert<ModelClass extends Model> implements Query, Queriable {
         }
 
         if (columns != null && values != null && columns.length != values.length) {
-            throw new IllegalStateException("The Insert of " + FlowManager.getTableName(table) + " when specifying" +
+            throw new IllegalStateException("The Insert of " + FlowManager.getTableName(getTable()) + " when specifying" +
                     "columns needs to have the same amount of values and columns");
         } else if (values == null) {
-            throw new IllegalStateException("The insert of " + FlowManager.getTableName(table) + " should have" +
+            throw new IllegalStateException("The insert of " + FlowManager.getTableName(getTable()) + " should have" +
                     "at least one value specified for the insert");
         }
 
@@ -235,38 +209,4 @@ public class Insert<ModelClass extends Model> implements Query, Queriable {
         return queryBuilder.getQuery();
     }
 
-    /**
-     * @return The table associated with this INSERT
-     */
-    public Class<ModelClass> getTable() {
-        return table;
-    }
-
-    @Override
-    public Cursor query() {
-        query(FlowManager.getDatabaseForTable(table).getWritableDatabase());
-        return null;
-    }
-
-    @Override
-    public Cursor query(DatabaseWrapper databaseWrapper) {
-        databaseWrapper.execSQL(getQuery());
-        return null;
-    }
-
-    @Override
-    public void execute() {
-        Cursor cursor = query();
-        if (cursor != null) {
-            cursor.close();
-        }
-    }
-
-    @Override
-    public void execute(DatabaseWrapper databaseWrapper) {
-        Cursor cursor = query(databaseWrapper);
-        if (cursor != null) {
-            cursor.close();
-        }
-    }
 }
