+++ dbflow/src/main/java/com/raizlabs/android/dbflow/sql/language/Set.java
@@ -1,9 +1,7 @@
 package com.raizlabs.android.dbflow.sql.language;
 
 import android.content.ContentValues;
-import android.database.Cursor;
 
-import com.raizlabs.android.dbflow.config.FlowManager;
 import com.raizlabs.android.dbflow.sql.Query;
 import com.raizlabs.android.dbflow.sql.QueryBuilder;
 import com.raizlabs.android.dbflow.sql.SqlUtils;
@@ -15,17 +13,15 @@ import com.raizlabs.android.dbflow.structure.database.DatabaseWrapper;
 /**
  * Description: Used to specify the SET part of an {@link com.raizlabs.android.dbflow.sql.language.Update} query.
  */
-public class Set<ModelClass extends Model> implements WhereBase<ModelClass>, Queriable, Transformable<ModelClass> {
+public class Set<ModelClass extends Model> extends BaseQueriable<ModelClass> implements WhereBase<ModelClass>, Queriable, Transformable<ModelClass> {
 
     private ConditionGroup conditionGroup;
 
     private Query update;
 
-    private Class<ModelClass> table;
-
     Set(Query update, Class<ModelClass> table) {
+        super(table);
         this.update = update;
-        this.table = table;
         conditionGroup = new ConditionGroup();
         conditionGroup.setAllCommaSeparated(true);
     }
@@ -127,40 +123,8 @@ public class Set<ModelClass extends Model> implements WhereBase<ModelClass>, Que
     }
 
     @Override
-    public Class<ModelClass> getTable() {
-        return table;
-    }
-
-    @Override
     public Query getQueryBuilderBase() {
         return update;
     }
 
-    @Override
-    public Cursor query() {
-        query(FlowManager.getDatabaseForTable(table).getWritableDatabase());
-        return null;
-    }
-
-    @Override
-    public Cursor query(DatabaseWrapper databaseWrapper) {
-        databaseWrapper.execSQL(getQuery());
-        return null;
-    }
-
-    @Override
-    public void execute() {
-        Cursor cursor = query();
-        if (cursor != null) {
-            cursor.close();
-        }
-    }
-
-    @Override
-    public void execute(DatabaseWrapper databaseWrapper) {
-        Cursor cursor = query(databaseWrapper);
-        if (cursor != null) {
-            cursor.close();
-        }
-    }
 }
